Data Generating Processes

AbstractDGP

DeepSimulatedMoments.AbstractDGP — Type

AbstractDGP{T<:AbstractFloat}

Abstract type for a DGP.

DeepSimulatedMoments.nfeatures — Function

nfeatures(d::AbstractDGP)

Number of features in the data.

Arguments

d::AbstractDGP: DGP to get the number of features from.

Returns

Int: Number of features.

DeepSimulatedMoments.nparams — Function

nparams(d::AbstractDGP)

Number of parameters in the DGP.

Arguments

d::AbstractDGP: DGP to get the number of parameters from.

Returns

Int: Number of parameters.

DeepSimulatedMoments.priordraw — Function

priordraw(d::AbstractDGP, S::Int)

Draw S parameter samples from the prior.

Arguments

d::AbstractDGP{T}: DGP to draw the parameters from.

S::Int: Number of samples to draw.

Returns

Matrix{T}: S samples of the parameters (dimension: S × nparams(d)).

DeepSimulatedMoments.generate — Function

generate(d::AbstractDGP, S::Int)

Generate S data and parameter samples from the DGP.

Arguments

d::AbstractDGP{T}: DGP to generate the data from.

S::Int: Number of samples to generate.

Returns

Tuple{Matrix{T}, Matrix{T}}: S samples of the data and parameters

(dimension: nfeatures(d) × S × N and S × nparams(d)).

generate(θ::AbstractVector{T}, d::AbstractDGP{T}, S::Int)

Generate S data samples from the DGP at parameters θ.

Arguments

θ::AbstractVector{T}: Parameters of DGP used to generate the data.

d::AbstractDGP{T}: DGP to generate the data from.

S::Int: Number of samples to generate.

Returns

Matrix{T}: S samples of the data (dimension: nfeatures(d) × S × N).

Predefined DGPs

DeepSimulatedMoments.MA2 — Type

MA2{T} <: AbstractDGP{T}

A simple MA(2) process. (T defaults to Float32)

Fields

N::Int: Number of observations in each sample.

DeepSimulatedMoments.Logit — Type

Logit{T} <: AbstractDGP{T}

A simple logistic regression model. (T defaults to Float32)

Fields

N::Int: Number of observations in each sample.

K::Int: Number of features in each sample.

DeepSimulatedMoments.GARCH — Type

GARCH{T} <: AbstractDGP{T}

A simple GARCH(1,1) process. (T defaults to Float32)

Fields

N::Int: Number of observations in each sample.

Neural Networks

Temporal Convolutional Networks

DeepSimulatedMoments.TemporalBlock — Function

TemporalBlock(chan_in, chan_out; dilation, kernel_size, residual, pad, dropout_rate)

Temporal block with chan_in input channels and chan_out output channels. Each block consists of two causal convolutional layers with kernel_size and dilation followed by batch normalization and dropout. If residual is true, a skip connection is added.

Arguments

chan_in::Int: Number of input channels.

chan_out::Int: Number of output channels.

Keyword arguments

dilation::Int: Kernel dilation.

kernel_size::Int: Size of the convolutional kernel.

residual::Bool: Whether to use residual connections.

pad::Function: Padding to use for the convolutional layers.

dropout_rate::AbstractFloat: Dropout rate to use for the convolutional layers.

Returns

Chain: Temporal block

DeepSimulatedMoments.TCN — Function

TCN(channels; kernel_size, dilation_factor, residual, pad, dropout_rate)

Temporal convolutional network (TCN) with length(channels) - 1 layers. Each layer is a TemporalBlock with channels[i] input channels and channels[i+1]

Arguments

channels::AbstractVector{Int}: Number of input and output channels for each layer.

kernel_size::Int: Size of the convolutional kernel.

dilation_factor::Int: Factor by which the dilation is increased for each layer. (default: 2)

residual::Bool: Whether to use residual connections. (default: true)

pad::Function: Padding to use for the convolutional layers. (default: SamePad())

dropout_rate::AbstractFloat: Dropout rate to use for the convolutional layers. (default: 0.)

Returns

Chain: TCN

Neural Network Utilities

DeepSimulatedMoments.tabular2conv — Function

tabular2conv(X::AbstractArray{T, 3}) where T

Transform a (K × S × T) array to a CNN format (1 × T × K × S array). K is the number of features, S is the number of samples, and T is the number of time steps / observations in each sample.

Arguments

X::AbstractArray{T, 3}: Array to transform.

Returns

Array{T, 4}: CNN format of X.

DeepSimulatedMoments.tabular2rnn — Function

tabular2rnn(X::AbstractArray{T, 3}) where T

Transform a (K × S × T) array to an RNN format (T-array of K × S matrices). K is the number of features, S is the number of samples, and T is the number of time steps / observations in each sample.

Arguments

X::AbstractArray{T, 3}: Array to transform.

Returns

Vector{Matrix{T}}: RNN format of X.