Data Generating Processes

AbstractDGP

DeepSimulatedMoments.AbstractDGP — Type

AbstractDGP{T<:AbstractFloat}

Abstract type for a DGP.

DeepSimulatedMoments.nfeatures — Function

nfeatures(d::AbstractDGP)

Number of features in the data.

Arguments

d::AbstractDGP: DGP to get the number of features from.

Returns

Int: Number of features.

DeepSimulatedMoments.nparams — Function

nparams(d::AbstractDGP)

Number of parameters in the DGP.

Arguments

d::AbstractDGP: DGP to get the number of parameters from.

Returns

Int: Number of parameters.

DeepSimulatedMoments.priordraw — Function

priordraw(d::AbstractDGP, S::Int)

Draw S parameter samples from the prior.

Arguments

d::AbstractDGP{T}: DGP to draw the parameters from.

S::Int: Number of samples to draw.

Returns

Matrix{T}: S samples of the parameters (dimension: S × nparams(d)).

DeepSimulatedMoments.generate — Function

generate(θ::AbstractVector{T}, d::AbstractDGP{T}, S::Int)

Generate S data samples from the DGP at parameters θ.

Arguments

θ::AbstractVector{T}: Parameters of DGP used to generate the data.

d::AbstractDGP{T}: DGP to generate the data from.

S::Int: Number of samples to generate.

Returns

Matrix{T}: S samples of the data (dimension: nfeatures(d) × S × N).

generate(dgp::AbstractDGP, net::MomentNetwork, nsamples::Int)

Generate data from a DGP and apply the data and parameter transforms from the moment network.

Arguments

dgp::AbstractDGP: The DGP to generate data from.

net::MomentNetwork: The moment network.

nsamples::Int: The number of samples to generate.

Returns

X: The transformed data.

Y: The transformed parameters.

DeepSimulatedMoments.datatransform — Function

datatransform(d::AbstractDGP, S::Int; dev=cpu)

Fit a ZScoreTransform data transform to the data generated from the DGP.

Arguments

d::AbstractDGP{T}: DGP to generate the data from.

S::Int: Number of samples to generate.

dev: Device to fit the transform on (default: cpu).

Returns

ZScoreTransform: Data transform.

DeepSimulatedMoments.likelihood — Function

likelihood(d::AbstractDGP, X::AbstractArray, θ::AbstractVector)

Compute the log-likelihood of the data X given the parameters θ. This function is irrelevant for DGPs with an intractable likelihood.

Arguments

d::AbstractDGP{T}: DGP to compute the log-likelihood for.

X::AbstractArray{T}: Data to compute the log-likelihood for.

θ::AbstractVector{T}: Parameters to compute the log-likelihood for.

Returns

T: Log-likelihood of the data given the parameters.

DeepSimulatedMoments.ErrorDistribution — Type

ErrorDistribution

A distribution for the error term in a AbstractDGP`.

Predefined DGPs

DeepSimulatedMoments.MA2 — Type

MA2{T} <: AbstractDGP{T}

A simple MA(2) process. (T defaults to Float32)

Fields

N::Int: Number of observations in each sample.

dist::ErrorDistribution: Distribution of the error term.

DeepSimulatedMoments.Logit — Type

Logit{T} <: AbstractDGP{T}

A simple logistic regression model. (T defaults to Float32)

Fields

N::Int: Number of observations in each sample.

K::Int: Number of features in each sample.

dist::ErrorDistribution: Distribution of the error term.

DeepSimulatedMoments.GARCH — Type

GARCH{T} <: AbstractDGP{T}

A simple GARCH(1,1) process. (T defaults to Float32)

Fields

N::Int: Number of observations in each sample.

dist::ErrorDistribution: Distribution of the error term.

Utilities

DeepSimulatedMoments.generate_files — Function

generatefiles(dgp::AbstractDGP, dir::String, nfiles::Int; batchsize::Int=1024, verbosity::Int=(nfiles ÷ 10), preprocessX::Function=identity, preprocessY::Function=identity )

Generate data files from a AbstractDGP and save it to a directory. Note that if the directory already contains files, the function will continue from the last file until the directory contains nfiles.

Arguments

dgp::AbstractDGP: The AbstractDGP to generate data from.

dir::String: The directory to save the data to.

nfiles::Int: The number of files to generate.

Keyword Arguments

batchsize::Int: The batch size (default: 1024).

verbosity::Int: The verbosity (default: nfiles ÷ 10).

preprocessX::Function: The preprocessing function for the data (default: identity).

preprocessY::Function: The preprocessing function for the parameters (default: identity).

DeepSimulatedMoments.load_from_file — Function

load_from_file(file::String)

Load pre-generated data from a file.

Arguments

file::String: The file to load the data from.

Moment Networks

HyperParameters

DeepSimulatedMoments.HyperParameters — Type

HyperParameters

Hyperparameters for training a MomentNetwork.

Fields

dev: The device to train on.

loss: The loss function.

batchsize: The batch size.

nsamples: The number of samples to generate.

epochs: The number of epochs to train for.

validation_size: The number of samples to use for validation.

validation_freq: The frequency with which to validate.

print_every: The frequency with which to print information about the training process.

MomentNetwork

DeepSimulatedMoments.MomentNetwork — Type

MomentNetwork

A neural network that estimates the parameters of a DGP.

Fields

model: The neural network.

optimizer: The optimizer.

best_model: The best model found during training.

hyperparameters: The hyperparameters.

preprocess: The preprocessing function.

parameter_transform: The parameter transform.

data_transform: The data transform.

DeepSimulatedMoments.train_network! — Function

train_network!(net::MomentNetwork, dgp::AbstractDGP; verbose::Bool=true)

Train a moment network on a DGP.

Arguments

net::MomentNetwork: The moment network.

dgp::AbstractDGP: The DGP to train on.

verbose::Bool: Whether to print information about the training process.

Returns

iterations: The iterations at which the losses were computed.

losses: The losses at each iteration.

DeepSimulatedMoments.apply_transforms — Function

apply_transforms(net::MomentNetwork, X, Y)

Apply the data and parameter transforms to the data and parameters, respectively.

Arguments

net::MomentNetwork: The moment network.

X: The data to apply the data transform to.

Y: The parameters to apply the parameter transform to.

Returns

X: The transformed data.

Y: The transformed parameters.

Neural Networks

Temporal Convolutional Networks

DeepSimulatedMoments.TemporalBlock — Function

TemporalBlock(chan_in, chan_out; dilation, kernel_size, residual, pad)

Temporal block with chan_in input channels and chan_out output channels. Each block consists of two causal convolutional layers with kernel_size and dilation followed by batch normalization. If residual is true, a skip connection is added.

Arguments

chan_in::Int: Number of input channels.

chan_out::Int: Number of output channels.

Keyword arguments

dilation::Int: Kernel dilation.

kernel_size::Int: Size of the convolutional kernel.

residual::Bool: Whether to use residual connections.

pad: Padding to use for the convolutional layers.

Returns

Chain: Temporal block

DeepSimulatedMoments.TCN — Function

TCN(channels; kernel_size, dilation_factor, residual, pad)

Temporal convolutional network (TCN) with length(channels) - 1 layers. Each layer is a TemporalBlock with channels[i] input channels and channels[i+1]

Arguments

channels::AbstractVector{Int}: Number of input and output channels for each layer.

kernel_size::Int: Size of the convolutional kernel.

dilation_factor::Int: Factor by which the dilation is increased for each layer. (default: 2)

residual::Bool: Whether to use residual connections. (default: true)

pad: Padding to use for the convolutional layers. (default: SamePad())

Returns

Chain: TCN

Utilities

DeepSimulatedMoments.build_tcn — Function

build_tcn(d::AbstractDGP; dilation_factor, kernel_size, channels, 
    summary_size, residual, pad)

Builds a TCN for a specified DGP.

Arguments

d::AbstractDGP: DGP to build the TCN for.

Keyword Arguments

nlayers::Int: Number of layers. (default: 0, computes the necessary layers using receptive_field_size)

dilation_factor::Int: Dilation factor. (default: 2)

kernel_size::Int: Size of the convolutional kernel. (default: 8)

channels::Int: Number of channels. (default: 32)

receptive_field_size::Int: Receptive field size. (default: 0, uses the length of the AbstractDGP instead)

summary_size::Int: Size of the summary convolution. (default: 10)

residual::Bool: Whether to use residual connections. (default: true)

pad: Padding function. (default: SamePad())

dev::Function: Device to use. (default: cpu)

Returns

Chain: TCN for the specified DGP.

DeepSimulatedMoments.receptive_field_size — Function

receptive_field_size(dilation, kernel_size, layers)

Computes the receptive field size for a specified dilation, kernel size, and number of layers.

Arguments

dilation::Int: Dilation factor.

kernel_size::Int: Size of the convolutional kernel.

layers::Int: Number of layers.

Returns

Int: Receptive field size.

DeepSimulatedMoments.necessary_layers — Function

necessary_layers(dilation, kernel_size, receptive_field)

Computes the minimum number of layers necessary to achieve a specified receptive field size.

Arguments

dilation::Int: Dilation factor.

kernel_size::Int: Size of the convolutional kernel.

receptive_field::Int: Desired receptive field size.

Keyword Arguments

use_ceil::Bool: Whether to use ceil(Int, x) for the output. (default: false)

use_floor::Bool: Whether to use floor(Int, x) for the output. (default: false)

Returns

Minimum number of layers (without use_ceil or use_floor, this is a Float64).

Recurrent Neural Networks

Utilities

General Neural Network Utilities

DeepSimulatedMoments.tabular2conv — Function

tabular2conv(X::AbstractArray{T, 3}) where T

Transform a (K × S × T) array to a CNN format (1 × T × K × S array). K is the number of features, S is the number of samples, and T is the number of time steps / observations in each sample.

Arguments

X::AbstractArray{T, 3}: Array to transform.

Returns

Array{T, 4}: CNN format of X.

DeepSimulatedMoments.tabular2rnn — Function

tabular2rnn(X::AbstractArray{T, 3}) where T

Transform a (K × S × T) array to an RNN format (T-array of K × S matrices). K is the number of features, S is the number of samples, and T is the number of time steps / observations in each sample.

Arguments

X::AbstractArray{T, 3}: Array to transform.

Returns

Vector{Matrix{T}}: RNN format of X.